backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
PyPDF2==3.0.1
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
chromadb==0.4.22
openai==1.6.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
tiktoken==0.5.2
pypdf==3.17.4
